``` refactor(task_manager): 将任务异常类提取到独立模块解决循环依赖
- 将 TaskSuccess、TaskFailed、TaskPauseForRateLimit 异常类从
  src/services/task_manager.py 移至 src/utils/task_exceptions.py
- 在原位置通过 re-export 保持对外接口不变，确保现有调用方无需修改
- task_profiler.py 改为从新模块导入异常类，避免 task_manager →
  src.db → utils → task_manager 的循环依赖问题
- 新的 task_exceptions 模块无外部依赖，可被多处安全导入
```
